src/requirements.txt
requests==2.28.1
tqdm==4.64.1
python-nmap